+++ src/gameLogic/action/Projectile.h
@@ -6,16 +6,17 @@
 class Projectile : public Attack
 {
     public:
-        Projectile(int x, int y, int dX, int dY, Sprite sprite, bool playerFriendly, int lifeTime);
+        Projectile(double x, double y, double dX, double dY, 
+                Sprite sprite, bool playerFriendly, int lifeTime);
 
         virtual void update();
 
         virtual void onObjectCollision(ObjectVector gameObjects);
         virtual void onTileCollision();
     private:
-        int dX, dY;
+        double dX, dY;
 
-        void move(int dX, int dY);
+        void move(double dX, double dY);
 };
 
 #endif
